+++ bndl/compute/context.py
@@ -1,5 +1,5 @@
 from bndl.compute.accumulate import Accumulator
-from bndl.compute.dense import sources as dense
+from bndl.compute.dense import arrays as dense
 from bndl.compute.broadcast import broadcast
 from bndl.compute.collections import DistributedCollection
 from bndl.compute.files import files