+++ docs/changelog.rst
@@ -5,6 +5,7 @@ Change log
 -----
  * Optimizations task/partition IO for shuffles
  * Bug in take_sample for tiny datasets
+ * Added random distributed numpy arrays
 
 0.3.2
 -----